+++ DirectX/Sound/Loader/WAV.h
@@ -19,9 +19,9 @@ public:
     ~WAV();
     virtual bool loadFromFile(WAVEFORMATEX* format, const std::string& fileName) override;
     //指定したサイズ分データを読み取る
-    virtual unsigned read(BYTE** buffer, unsigned size) override;
+    virtual long read(BYTE** buffer, unsigned size) override;
     //オフセット分シーク
-    virtual unsigned seek(int offset, Seek seek = Seek::CURRENT) override;
+    virtual long seek(int offset, Seek seek = Seek::CURRENT) override;
     //バッファ部分の先頭までシーク
     virtual void seekBegin() override;
     //波形部分の大きさ
